Fix argument type in ProxyCommitData::updateSSTagCost

+++ fdbserver/include/fdbserver/ProxyCommitData.actor.h
@@ -303,7 +303,7 @@ struct ProxyCommitData {
 		latencyBandConfig = newLatencyBandConfig;
 	}
 
-	void updateSSTagCost(const UID& id, const TagSet& tagSet, MutationRef m, int cost) {
+	void updateSSTagCost(const UID& id, const TagSet& tagSet, MutationRef m, uint64_t cost) {
 		auto [it, _] = ssTrTagCommitCost.try_emplace(id, TransactionTagMap<TransactionCommitCostEstimation>());
 
 		for (auto& tag : tagSet) {
